Restuctured,added view submitted invoice page
